tests: Move MiscMessageTest to test_audit_log.py.
